pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["src/app"]

[project]
name = "app-qg-api"
version = "0.1.0"
description = "FastAPI starter with async connectors"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.0",
    "uvicorn[standard]>=0.38.0",
    "pydantic-settings>=2.12.0",
    "structlog>=25.5.0",
    "motor>=3.7.1",
    "sqlalchemy[asyncio]>=2.0.44",
    "geoalchemy2>=0.15.2",
    "asyncpg>=0.31.0",
    "aio-pika>=9.5.8",
    "httpx>=0.28.1",
    "pyjwt[crypto]>=2.9.0",
]

[dependency-groups]
dev = [
    "ruff>=0.7.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.28.1",
]

[tool.ruff]
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = [
    "B008",
    "E501",
    "B904",
    "B006",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["src"]
testpaths = ["tests"]

[tool.uv.workspace]
members = [
    ".",
]

[tool.uv.sources]
app-qg-api = { workspace = true, editable = true }
